Initial commit of MPL license

+++ Wurst Client/src/org/darkstorm/minecraft/gui/theme/wurst/WurstSliderUI.java
@@ -1,3 +1,10 @@
+/*
+ * Copyright © 2014 - 2015 | Alexander01998 | All rights reserved.
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
 package org.darkstorm.minecraft.gui.theme.wurst;
 
 import static org.lwjgl.opengl.GL11.GL_BLEND;
@@ -26,19 +33,22 @@ import org.darkstorm.minecraft.gui.theme.AbstractComponentUI;
 import org.darkstorm.minecraft.gui.util.RenderUtil;
 import org.lwjgl.input.Mouse;
 
-public class WurstSliderUI extends AbstractComponentUI<Slider> {
+public class WurstSliderUI extends AbstractComponentUI<Slider>
+{
 	private WurstTheme theme;
-
-	public WurstSliderUI(WurstTheme theme) {
+	
+	public WurstSliderUI(WurstTheme theme)
+	{
 		super(Slider.class);
 		this.theme = theme;
-
+		
 		foreground = Color.LIGHT_GRAY;
 		background = new Color(128, 128, 128, 128 + 128 / 2);
 	}
-
+	
 	@Override
-	protected void renderComponent(Slider component) {
+	protected void renderComponent(Slider component)
+	{
 		translateComponent(component, false);
 		glEnable(GL_BLEND);
 		glDisable(GL_CULL_FACE);
@@ -47,26 +57,28 @@ public class WurstSliderUI extends AbstractComponentUI<Slider> {
 		FontRenderer fontRenderer = theme.getFontRenderer();
 		fontRenderer.drawString(component.getText(), 0, 0, RenderUtil.toRGBA(component.getForegroundColor()));
 		String content = null;
-		switch(component.getValueDisplay()) {
-		case DECIMAL:
-			content = Double.toString((double)((long)(Math.round(component.getValue() / component.getIncrement()) * 1000000) * (long)(component.getIncrement() * 1000000)) / 1000000 / 1000000);
-			break;
-		case INTEGER:
-			content = String.format("%,d", Long.valueOf(Math.round(component.getValue())));
-			break;
-		case PERCENTAGE:
-			int percent = (int) Math.round((component.getValue() - component.getMinimumValue()) / (component.getMaximumValue() - component.getMinimumValue()) * 100D);
-			content = String.format("%d%%", percent);
-		default:
+		switch(component.getValueDisplay())
+		{
+			case DECIMAL:
+				content = Double.toString((double)(Math.round(component.getValue() / component.getIncrement()) * 1000000 * (long)(component.getIncrement() * 1000000)) / 1000000 / 1000000);
+				break;
+			case INTEGER:
+				content = String.format("%,d", Long.valueOf(Math.round(component.getValue())));
+				break;
+			case PERCENTAGE:
+				int percent = (int)Math.round((component.getValue() - component.getMinimumValue()) / (component.getMaximumValue() - component.getMinimumValue()) * 100D);
+				content = String.format("%d%%", percent);
+			default:
 		}
-		if(content != null) {
+		if(content != null)
+		{
 			String suffix = component.getContentSuffix();
 			if(suffix != null && !suffix.trim().isEmpty())
 				content = content.concat(" ").concat(suffix);
 			fontRenderer.drawString(content, component.getWidth() - fontRenderer.getStringWidth(content), 0, RenderUtil.toRGBA(component.getForegroundColor()));
 		}
 		glDisable(GL_TEXTURE_2D);
-
+		
 		glColor4f(0.0f, 0.0f, 0.0f, 1.0f);
 		glLineWidth(1.0f);
 		glBegin(GL_LINE_LOOP);
@@ -111,30 +123,36 @@ public class WurstSliderUI extends AbstractComponentUI<Slider> {
 		glDisable(GL_BLEND);
 		translateComponent(component, true);
 	}
-
+	
 	@Override
-	protected Dimension getDefaultComponentSize(Slider component) {
+	protected Dimension getDefaultComponentSize(Slider component)
+	{
 		return new Dimension(106, 8 + theme.getFontRenderer().FONT_HEIGHT);
 	}
-
+	
 	@Override
-	protected Rectangle[] getInteractableComponentRegions(Slider component) {
-		return new Rectangle[] { new Rectangle(0, theme.getFontRenderer().FONT_HEIGHT + 2, component.getWidth(), component.getHeight() - theme.getFontRenderer().FONT_HEIGHT) };
+	protected Rectangle[] getInteractableComponentRegions(Slider component)
+	{
+		return new Rectangle[]{new Rectangle(0, theme.getFontRenderer().FONT_HEIGHT + 2, component.getWidth(), component.getHeight() - theme.getFontRenderer().FONT_HEIGHT)};
 	}
-
+	
 	@Override
-	protected void handleComponentInteraction(Slider component, Point location, int button) {
+	protected void handleComponentInteraction(Slider component, Point location, int button)
+	{
 		if(getInteractableComponentRegions(component)[0].contains(location) && button == 0)
 			if(Mouse.isButtonDown(button) && !component.isValueChanging())
 				component.setValueChanging(true);
 			else if(!Mouse.isButtonDown(button) && component.isValueChanging())
 				component.setValueChanging(false);
 	}
-
+	
 	@Override
-	protected void handleComponentUpdate(Slider component) {
-		if(component.isValueChanging()) {
-			if(!Mouse.isButtonDown(0)) {
+	protected void handleComponentUpdate(Slider component)
+	{
+		if(component.isValueChanging())
+		{
+			if(!Mouse.isButtonDown(0))
+			{
 				component.setValueChanging(false);
 				return;
 			}
@@ -142,8 +160,8 @@ public class WurstSliderUI extends AbstractComponentUI<Slider> {
 			Container parent = component.getParent();
 			if(parent != null)
 				mouse.translate(-parent.getX(), -parent.getY());
-			double percent = (double) (mouse.x - 4) / (double) (component.getWidth() - 6);
-			double value = component.getMinimumValue() + (percent * (component.getMaximumValue() - component.getMinimumValue()));
+			double percent = (double)(mouse.x - 4) / (double)(component.getWidth() - 6);
+			double value = component.getMinimumValue() + percent * (component.getMaximumValue() - component.getMinimumValue());
 			component.setValue(value);
 		}
 	}
