+++ Wurst Client/src/tk/wurst_client/utils/BlockUtils.java
@@ -1,3 +1,10 @@
+/*
+ * Copyright © 2014 - 2015 | Alexander01998 | All rights reserved.
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
 package tk.wurst_client.utils;
 
 import net.minecraft.client.Minecraft;
@@ -9,34 +16,34 @@ public class BlockUtils
 {
 	public static void faceBlockClient(BlockPos blockPos)
 	{
-        double diffX = blockPos.getX() + 0.5 - Minecraft.getMinecraft().thePlayer.posX;
-        double diffY = blockPos.getY() + 0.5 - (Minecraft.getMinecraft().thePlayer.posY + (double)Minecraft.getMinecraft().thePlayer.getEyeHeight());
-        double diffZ = blockPos.getZ() + 0.5 - Minecraft.getMinecraft().thePlayer.posZ;
-        double dist = (double)MathHelper.sqrt_double(diffX * diffX + diffZ * diffZ);
-        float yaw = (float)(Math.atan2(diffZ, diffX) * 180.0D / Math.PI) - 90.0F;
-        float pitch = (float)(-(Math.atan2(diffY, dist) * 180.0D / Math.PI));
-        Minecraft.getMinecraft().thePlayer.rotationYaw = Minecraft.getMinecraft().thePlayer.rotationYaw + MathHelper.wrapAngleTo180_float(yaw - Minecraft.getMinecraft().thePlayer.rotationYaw);
-        Minecraft.getMinecraft().thePlayer.rotationPitch = Minecraft.getMinecraft().thePlayer.rotationPitch + MathHelper.wrapAngleTo180_float(pitch - Minecraft.getMinecraft().thePlayer.rotationPitch);
+		double diffX = blockPos.getX() + 0.5 - Minecraft.getMinecraft().thePlayer.posX;
+		double diffY = blockPos.getY() + 0.5 - (Minecraft.getMinecraft().thePlayer.posY + Minecraft.getMinecraft().thePlayer.getEyeHeight());
+		double diffZ = blockPos.getZ() + 0.5 - Minecraft.getMinecraft().thePlayer.posZ;
+		double dist = MathHelper.sqrt_double(diffX * diffX + diffZ * diffZ);
+		float yaw = (float)(Math.atan2(diffZ, diffX) * 180.0D / Math.PI) - 90.0F;
+		float pitch = (float)-(Math.atan2(diffY, dist) * 180.0D / Math.PI);
+		Minecraft.getMinecraft().thePlayer.rotationYaw = Minecraft.getMinecraft().thePlayer.rotationYaw + MathHelper.wrapAngleTo180_float(yaw - Minecraft.getMinecraft().thePlayer.rotationYaw);
+		Minecraft.getMinecraft().thePlayer.rotationPitch = Minecraft.getMinecraft().thePlayer.rotationPitch + MathHelper.wrapAngleTo180_float(pitch - Minecraft.getMinecraft().thePlayer.rotationPitch);
 	}
-	
+
 	public static void faceBlockPacket(BlockPos blockPos)
 	{
-        double diffX = blockPos.getX() + 0.5 - Minecraft.getMinecraft().thePlayer.posX;
-        double diffY = blockPos.getY() + 0.5 - (Minecraft.getMinecraft().thePlayer.posY + (double)Minecraft.getMinecraft().thePlayer.getEyeHeight());
-        double diffZ = blockPos.getZ() + 0.5 - Minecraft.getMinecraft().thePlayer.posZ;
-        double dist = (double)MathHelper.sqrt_double(diffX * diffX + diffZ * diffZ);
-        float yaw = (float)(Math.atan2(diffZ, diffX) * 180.0D / Math.PI) - 90.0F;
-        float pitch = (float)(-(Math.atan2(diffY, dist) * 180.0D / Math.PI));
-        Minecraft.getMinecraft().thePlayer.sendQueue.addToSendQueue(new C03PacketPlayer.C05PacketPlayerLook
-        (
-        	Minecraft.getMinecraft().thePlayer.rotationYaw + MathHelper.wrapAngleTo180_float(yaw - Minecraft.getMinecraft().thePlayer.rotationYaw),
-        	Minecraft.getMinecraft().thePlayer.rotationPitch + MathHelper.wrapAngleTo180_float(pitch - Minecraft.getMinecraft().thePlayer.rotationPitch),
-        	Minecraft.getMinecraft().thePlayer.onGround
-        ));
-    }
-	
+		double diffX = blockPos.getX() + 0.5 - Minecraft.getMinecraft().thePlayer.posX;
+		double diffY = blockPos.getY() + 0.5 - (Minecraft.getMinecraft().thePlayer.posY + Minecraft.getMinecraft().thePlayer.getEyeHeight());
+		double diffZ = blockPos.getZ() + 0.5 - Minecraft.getMinecraft().thePlayer.posZ;
+		double dist = MathHelper.sqrt_double(diffX * diffX + diffZ * diffZ);
+		float yaw = (float)(Math.atan2(diffZ, diffX) * 180.0D / Math.PI) - 90.0F;
+		float pitch = (float)-(Math.atan2(diffY, dist) * 180.0D / Math.PI);
+		Minecraft.getMinecraft().thePlayer.sendQueue.addToSendQueue(new C03PacketPlayer.C05PacketPlayerLook
+			(
+				Minecraft.getMinecraft().thePlayer.rotationYaw + MathHelper.wrapAngleTo180_float(yaw - Minecraft.getMinecraft().thePlayer.rotationYaw),
+				Minecraft.getMinecraft().thePlayer.rotationPitch + MathHelper.wrapAngleTo180_float(pitch - Minecraft.getMinecraft().thePlayer.rotationPitch),
+				Minecraft.getMinecraft().thePlayer.onGround
+			));
+	}
+
 	public static float getBlockDistance(float xDiff, float yDiff, float zDiff)
 	{
-		return MathHelper.sqrt_float((xDiff - 0.5F) * (xDiff - 0.5F) + (yDiff - 0.5F) * (yDiff - 0.5F) + (zDiff - 0.5F) *( zDiff - 0.5F));
+		return MathHelper.sqrt_float((xDiff - 0.5F) * (xDiff - 0.5F) + (yDiff - 0.5F) * (yDiff - 0.5F) + (zDiff - 0.5F) * (zDiff - 0.5F));
 	}
 }
